Use smart layout that picks correct header
