+++ frontend/src/routes/(unauthenticated)/+layout.svelte
@@ -1,9 +1,7 @@
 <script>
-  import { AppBar, Content } from '$lib/layout';
+  import { Layout } from '$lib/layout';
 </script>
 
-<AppBar user={undefined} />
-
-<Content>
+<Layout>
   <slot />
-</Content>
+</Layout>
